requirement.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
CacheControl==0.12.10
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
Django==4.0.3
django-filter==21.1
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
djangorestframework-datatables==0.7.0
fcm-django==1.0.11
firebase-admin==5.2.0
google-api-core==2.7.1
google-api-python-client==2.43.0
google-auth==2.6.2
google-auth-httplib2==0.1.0
google-cloud-core==2.2.3
google-cloud-firestore==2.4.0
google-cloud-storage==2.2.1
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpcio==1.44.0
grpcio-status==1.44.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.3
Markdown==3.3.6
msgpack==1.0.3
Pillow==9.1.0
proto-plus==1.20.3
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyMySQL==1.0.2
pyparsing==3.0.7
pytz==2022.1
requests==2.27.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
stripe==2.70.0
uritemplate==4.1.1
urllib3==1.26.9
zipp==3.8.0
